Fix worker file handles being held during test runtime
`tempfile.mkstemp` is creating a file handle which is never closed. This doesn't matter on linux, but when running these tests on Windows machines, will cause an `OSError` when trying to remove the files.

This fixes the problem by using `mktemp` instead, which only returns a name, and then using `Path.touch()` to create the file but not keep an open handle.